Extract modal_osworld_direct shared code into mantis_agent.modal_runtime
The 1900-line modal_osworld_direct.py was acting as both a Modal entry-point
and a de-facto shared library: 5 sibling files (modal_web_tasks.py,
modal_test_gemma4_cua.py, benchmarks/osworld_chrome.py,
benchmarks/osworld_multiapp.py, benchmarks/visualwebarena.py) imported
image, vol, GEMMA4_MODEL, GGUF_CONFIGS, download_model, start_llama_server,
and run_osworld_impl from it via top-level imports backed by sys.path
hacks.

Refactor:
- src/mantis_agent/modal_runtime.py — new module owning image, vol, model
  config, llama-server bringup, OSWorld helpers, and run_osworld_impl.
  No modal.App at module scope — each entry-point owns its app namespace.
- modal_osworld_direct.py — now a 50-line thin Modal entry-point that
  imports from mantis_agent.modal_runtime and defines its own app +
  @app.function + @app.local_entrypoint.
- 5 callers updated to import from mantis_agent.modal_runtime.
- Dropped sys.path.insert(parent_root) shims from benchmarks/*.
- image dropped .add_local_python_source("modal_osworld_direct"); the
  package add covers the new module.

Validated: modal run --help works for all 6 affected entry-points,
105 tests pass.
